Fix custom sorting for Reports
