+++ frontend/app/components/Reports/ReportsCardView.js
@@ -53,4 +53,4 @@ class ReportsCardView extends Component {
   }
 }
 
-export { ReportsCardView }
+export default ReportsCardView
